FIX MULTIPLE ERRORS SHOWN IN PHP LOG IF ARTICLE HAD EMPTY FIELD
If article does not have any data in custom field, the module
is still displayed. It is empty, but there are a log of php errors
in server log files. This commit fixes that.
It checks if article does have a custom field with a value stored.
if that is not empty, it display the module and its content.

closes #4