FIX: Not showing images inserted on Joomla 4
If you try to insert image with Joomla 4, it will add image path and
some aditional data. This plugin strips that extra data and exposes just
the relative image path.

I will try to create a less complex solution that will cover more use
cases as soon as more documentation, examples and tutorials became
available. And when I find an explanation on why this behaivour exists in
Joomla 4.