CI: Fix for update server
While doing automatic releases, the CI created new entry on update server,
but with the release of J!4 some tags were deprecated and some were
changed. This commit contains fixes for them. It also fixes the bug where
the CI script used tag name instead of version number in the update
section.

This commit also introduce automatic edit of main xml file to update
version and date of the release.

This commit manually edit the old update server file so it is up to date
with the correct syntax.